+++ python_modules/dagster/dagster_tests/cli_tests/workspace_tests/test_repository_load.py
@@ -209,3 +209,18 @@ def test_local_directory_module():
 )
 def test_local_directory_file(cli_args):
     successfully_load_repository_via_cli(cli_args)
+
+
+def test_dagster_definitions():
+    cli_args = ["-f", file_relative_path(__file__, "definitions_test_cases/defs_file.py")]
+
+    executed = {}
+
+    def the_assert(external_repo: ExternalRepository):
+        assert external_repo.name == "__repository__"
+        assert len(external_repo.get_external_asset_nodes()) == 1
+        executed["yes"] = True
+
+    assert successfully_load_repository_via_cli(cli_args, the_assert)
+
+    assert executed["yes"]
